Move symbols to root
