+++ src/index.js
@@ -2,6 +2,5 @@ import ReduxWPAPI from './ReduxWPAPI';
 
 export default ReduxWPAPI;
 export * as RequestStatus from './constants/requestStatus';
-export * as Symbols from './constants/symbols';
 export callAPI from './actions/callAPI';
 export { selectQuery } from './selectors';
